try different approach for porting pyodide.js
